+++ tools/jit/gen_unboxing_wrappers.py
@@ -20,10 +20,11 @@ torch/csrc/jit/generated/
 
 import argparse
 import re
+import yaml
 from itertools import groupby
 from ..autograd.gen_autograd import load_aten_declarations
 from ..autograd.gen_autograd import RETURNS_VIEWS_OF_INPUT
-from ..autograd.utils import CodeTemplate, write, is_out_variant, signature_without_args
+from ..autograd.utils import CodeTemplate, write, is_out_variant, signature_without_args, YamlLoader
 
 # JIT has a type system of
 # Scalar = int | float | bool # int is the largest int (int64_t),
@@ -272,12 +273,17 @@ def is_backward_op(decl):
 def argument_order(decl):
     return decl.get('jit_argument_order') or list(range(len(decl['arguments'])))
 
+def load_op_list(path):
+    with open(path, 'r') as f:
+        op_list = yaml.load(f, Loader=YamlLoader)
+    return op_list
 
 def gen_unboxing_wrappers(
     declarations,
     out,
     template_path,
     disable_autograd=False,
+    selected_op_list_path=None,
     selected_op_list=None,
     force_schema_registration=False,
 ):
@@ -464,6 +470,9 @@ def gen_unboxing_wrappers(
             reorder_out_args(decl)
 
     jit_decls.extend(additional_jit_decls)
+    if not selected_op_list:
+        selected_op_list = []
+    selected_op_list += load_op_list(selected_op_list_path) if selected_op_list_path else []
     jit_decls = filter_decls(jit_decls, disable_autograd, selected_op_list, force_schema_registration)
 
     # generation is deterministic
